azure_foundary_hosted_agents-main/coaching-skill-creator/requirements.txt
azure-ai-agentserver-agentframework>=1.0.0b17
python-frontmatter>=1.1.0
httpx>=0.27.0
